An emulator.
We need a way to test the compiler, which will emit assembly. This is a
really quick-and-dirty assembly emulator, which should let us test
things.

It will need to be modified to fit our testing framework, but for now it
should give us what we need to run a file through the emulator, and then
emit (a) the current machine state and (b) return the value of RAM
location zero.

We have informally decided that the result of any and all computation
should always end up in RAM 0. (Or, formally?)