Interpreter is probably working...
This is not yet "tested."

What I've done is work through the ten.hack code, updating the
emulator with better debugging, and (along the way) doing things that I
know need doing to make the code more robust/work. Put another way,
there were one or two things that were clearly naive and/or wrong, which
I've fixed.

I'll now try some specific tests, and see if JNE is working correctly. I
believe, though, that the ASM is laid out incorrectly at this point. So,
it will be time to revisit the code tomorrow and see what comes.