Fixes.
Jumps were not quite right. In the case of

0;JMP

I was jumping to zero. Really, the computation should happen and then be
ignored. (That is, the computation says what the ALU should output, not
where it should go.) Hence, the 0 just means "emit a zero from the ALU,
and don't store it." The JMP is an unconditional jump to the contents of
the A register.

Hence, this should do the trick.