Added provide statements and new "driver".
So, I recmomend using "run-all-three.rkt" for exploration. It is because
I am too lazy to explore things using multiple tabs...

(all-three "filename.420")

is the way to go.

It will show you the result of running the interpreter on the input
file, then run the compiler driver, and then run the emulator on the
resulting assembly.

I created two tests:

if-result-8.420
if-result-42.420

based on Trent's exploration.

Looking at the assembly from if-result-8.420, it looks... wrong.

@0
D=A
@num0
M=D
@num0
D=M
@TEST13284
M=D
@TEST13284
D=M
@sym0
M=D
@sym0
D=M
@FALSE-LABEL13282
D;JNE
@8
D=A
@num1
M=D
@ENDIF-LABEL13283
0;JMP
(FALSE-LABEL13282)
@42
D=A
@num2
M=D
@ENDIF-LABEL13283
0;JMP
(ENDIF-LABEL13283)
@0
M=D

Notice how zero is loaded into memory at location @num0. That's OK.
Then... it loads the location of label @TEST13284, and does something
with it... and, then shuffles the value located at @sym0 back-=and-forth
from memory.

I have no idea if we end up with a zero in D when we get to the D;JNE
instruction.

So, it seems like some focus is needed around that part of the
expansion. I think. Perhaps.

In other words... Trent, you're on the right track.

M