Emulator was incorrect.
So, I'll be happy to show you the emulator on Tuesday.

It was wrong.

Specifically, the emulator was building a jump structure that looked
like this:

(jump type target)

and then using the jump target for the comparison. So, if we were
jumping to line 22 of the code, the JNE was asking if line 22 was zero.
This, of course, is silly.

I've fixed the structure so it looks like

(jump type value target)

and it asks if the value (which is the result of the computation) is
zero. In the case of JNE, if it is not equal to zero, the code jumps to
the target.

This fixed the emulator. It still bears testing, but it seems like our
problem was in in the emulator, not in the compiler.

Good work on all of you, poor work on me. It was Trent's debugging this
afternoon that lead me to look more closely, because his tests were not
that complex, and... seemed like they should be working. That implied to
me that things were possibly not right somewhere else. (Honestly, I was
also suspicious of the compiler... we do a lot of shuffling of values.
However, it may be that *I* was wrong and *y'all* were right.)

Cheers,
M